refactor: move most functions from main program to their proper modules
